+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/codewhisperer/language/languages/CodeWhispererJsx.kt
@@ -5,13 +5,12 @@ package software.aws.toolkits.jetbrains.services.codewhisperer.language.language
 
 import software.aws.toolkits.jetbrains.services.codewhisperer.language.CodeWhispererProgrammingLanguage
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.FileCrawler
-import software.aws.toolkits.jetbrains.services.codewhisperer.util.JavascriptCodeWhispererFileCrawler
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.NoOpFileCrawler
 import software.aws.toolkits.telemetry.CodewhispererLanguage
 
 class CodeWhispererJsx private constructor() : CodeWhispererProgrammingLanguage() {
     override val languageId: String = ID
-    override val fileCrawler: FileCrawler = FileCrawler.EP_NAME.findFirstSafe { it is JavascriptCodeWhispererFileCrawler }
+    override val fileCrawler: FileCrawler = FileCrawler.EP_NAME.findFirstSafe { it.id == "javascript" }
         ?: NoOpFileCrawler()
 
     override fun toTelemetryType(): CodewhispererLanguage = CodewhispererLanguage.Jsx
