+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/codewhisperer/language/languages/CodeWhispererTsx.kt
@@ -6,12 +6,11 @@ package software.aws.toolkits.jetbrains.services.codewhisperer.language.language
 import software.aws.toolkits.jetbrains.services.codewhisperer.language.CodeWhispererProgrammingLanguage
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.FileCrawler
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.NoOpFileCrawler
-import software.aws.toolkits.jetbrains.services.codewhisperer.util.TypescriptCodeWhispererFileCrawler
 import software.aws.toolkits.telemetry.CodewhispererLanguage
 
 class CodeWhispererTsx private constructor() : CodeWhispererProgrammingLanguage() {
     override val languageId: String = ID
-    override val fileCrawler: FileCrawler = FileCrawler.EP_NAME.findFirstSafe { it is TypescriptCodeWhispererFileCrawler }
+    override val fileCrawler: FileCrawler = FileCrawler.EP_NAME.findFirstSafe { it.id == "typescript" }
         ?: NoOpFileCrawler()
 
     override fun toTelemetryType(): CodewhispererLanguage = CodewhispererLanguage.Tsx
