--- jetbrains-ultimate/src/software/aws/toolkits/jetbrains/services/codewhisperer/languages/JavascriptFileCrawler.kt
@@ -1,23 +1,0 @@
-// Copyright 2023 Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package software.aws.toolkits.jetbrains.services.codewhisperer.languages
-
-import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.psi.PsiFile
-import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererFileCrawler
-
-class JavascriptCodeWhispererFileCrawler : CodeWhispererFileCrawler() {
-    override val fileExtension: String = "js"
-    override val dialects: Set<String> = setOf("js", "jsx")
-    override val testFileNamingPatterns: List<Regex> = listOf(
-        Regex("""^(.+)\.(?i:t)est(\.js|\.jsx)$"""),
-        Regex("""^(.+)\.(?i:s)pec(\.js|\.jsx)$""")
-    )
-
-    override suspend fun listFilesImported(psiFile: PsiFile): List<VirtualFile> = emptyList()
-
-    override fun findSourceFileByName(psiFile: PsiFile): VirtualFile? = null
-
-    override fun findSourceFileByContent(psiFile: PsiFile): VirtualFile? = null
-}
