+++ jetbrains-ultimate/tst/software/aws/toolkits/jetbrains/services/codewhisperer/languages/JavascriptCodeWhispererFileCrawlerTest.kt
@@ -12,7 +12,7 @@ import org.junit.Test
 import software.aws.toolkits.jetbrains.services.codewhisperer.util.CodeWhispererFileCrawler
 import software.aws.toolkits.jetbrains.utils.rules.CodeInsightTestFixtureRule
 
-class TypescriptCodeWhispererFileCrawlerTest {
+class JavascriptCodeWhispererFileCrawlerTest {
     @JvmField
     @Rule
     val projectRule: CodeInsightTestFixtureRule = CodeInsightTestFixtureRule()
@@ -24,7 +24,7 @@ class TypescriptCodeWhispererFileCrawlerTest {
 
     @Before
     fun setup() {
-        sut = TypescriptCodeWhispererFileCrawler()
+        sut = JavascriptCodeWhispererFileCrawler()
 
         project = projectRule.project
         fixture = projectRule.fixture
@@ -32,60 +32,60 @@ class TypescriptCodeWhispererFileCrawlerTest {
 
     @Test
     fun `isTest - should return false`() {
-        val file1 = fixture.addFileToProject("src/utils/foo.ts", "")
+        val file1 = fixture.addFileToProject("src/utils/foo.js", "")
         assertThat(sut.isTestFile(file1.virtualFile, project)).isFalse
 
-        val file2 = fixture.addFileToProject("src/controler/bar.tsx", "")
+        val file2 = fixture.addFileToProject("src/controler/bar.jsx", "")
         assertThat(sut.isTestFile(file2.virtualFile, project)).isFalse
 
-        val file3 = fixture.addFileToProject("main.ts", "")
+        val file3 = fixture.addFileToProject("main.js", "")
         assertThat(sut.isTestFile(file3.virtualFile, project)).isFalse
 
-        val file4 = fixture.addFileToProject("component/dto/boo.tsx", "")
+        val file4 = fixture.addFileToProject("component/dto/boo.jsx", "")
         assertThat(sut.isTestFile(file4.virtualFile, project)).isFalse
     }
 
     @Test
     fun `isTest - should return true`() {
-        val file1 = fixture.addFileToProject("tst/components/foo.test.ts", "")
+        val file1 = fixture.addFileToProject("tst/components/foo.test.js", "")
         assertThat(sut.isTestFile(file1.virtualFile, project)).isTrue
 
-        val file2 = fixture.addFileToProject("test/components/foo.spec.ts", "")
+        val file2 = fixture.addFileToProject("test/components/foo.spec.js", "")
         assertThat(sut.isTestFile(file2.virtualFile, project)).isTrue
 
-        val file3 = fixture.addFileToProject("tests/components/foo.test.tsx", "")
+        val file3 = fixture.addFileToProject("tests/components/foo.test.jsx", "")
         assertThat(sut.isTestFile(file3.virtualFile, project)).isTrue
 
-        val file4 = fixture.addFileToProject("foo.spec.tsx", "")
+        val file4 = fixture.addFileToProject("foo.spec.jsx", "")
         assertThat(sut.isTestFile(file4.virtualFile, project)).isTrue
 
-        val file5 = fixture.addFileToProject("foo.test.ts", "")
+        val file5 = fixture.addFileToProject("foo.test.js", "")
         assertThat(sut.isTestFile(file5.virtualFile, project)).isTrue
 
-        val file6 = fixture.addFileToProject("src/tst/services/fooService.test.ts", "")
+        val file6 = fixture.addFileToProject("src/tst/services/fooService.test.js", "")
         assertThat(sut.isTestFile(file6.virtualFile, project)).isTrue
 
-        val file7 = fixture.addFileToProject("tests/services/barService.spec.tsx", "")
+        val file7 = fixture.addFileToProject("tests/services/barService.spec.jsx", "")
         assertThat(sut.isTestFile(file7.virtualFile, project)).isTrue
 
-        val file8 = fixture.addFileToProject("foo.Test.ts", "")
+        val file8 = fixture.addFileToProject("foo.Test.js", "")
         assertThat(sut.isTestFile(file8.virtualFile, project)).isTrue
 
-        val file9 = fixture.addFileToProject("foo.Spec.ts", "")
+        val file9 = fixture.addFileToProject("foo.Spec.js", "")
         assertThat(sut.isTestFile(file9.virtualFile, project)).isTrue
     }
 
     @Test
-    fun `guessSourceFileName typescript`() {
-        assertThat(sut.guessSourceFileName("fooBar.test.ts")).isEqualTo("fooBar.ts")
-        assertThat(sut.guessSourceFileName("fooBar.spec.ts")).isEqualTo("fooBar.ts")
-        assertThat(sut.guessSourceFileName("fooBarNoIdea.ts")).isNull()
+    fun `guessSourceFileName javascript`() {
+        assertThat(sut.guessSourceFileName("fooBar.test.js")).isEqualTo("fooBar.js")
+        assertThat(sut.guessSourceFileName("fooBar.spec.js")).isEqualTo("fooBar.js")
+        assertThat(sut.guessSourceFileName("fooBarNoIdea.js")).isNull()
     }
 
     @Test
-    fun `guessSourceFileName tsx`() {
-        assertThat(sut.guessSourceFileName("fooBar.test.tsx")).isEqualTo("fooBar.tsx")
-        assertThat(sut.guessSourceFileName("fooBar.spec.tsx")).isEqualTo("fooBar.tsx")
-        assertThat(sut.guessSourceFileName("fooBarNoIdea.tsx")).isNull()
+    fun `guessSourceFileName jsx`() {
+        assertThat(sut.guessSourceFileName("fooBar.test.jsx")).isEqualTo("fooBar.jsx")
+        assertThat(sut.guessSourceFileName("fooBar.spec.jsx")).isEqualTo("fooBar.jsx")
+        assertThat(sut.guessSourceFileName("fooBarNoIdea.jsx")).isNull()
     }
 }
